backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
openai>=1.0.0
httpx>=0.27.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
# manim-physics does not yet support Python 3.13+; skipped (gracefully handled in code)
# manim-physics>=0.2.4
pymunk>=6.0.0
scipy>=1.13.0
matplotlib>=3.9.0
